% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bogo_sort.R
\name{bogo_sort}
\alias{bogo_sort}
\title{bogo_sort}
\usage{
bogo_sort(lst, ascending = TRUE)
}
\arguments{
\item{lst}{An iterable object.}

\item{ascending}{Boolean specifying whether to sort in ascending or descending order.}
}
\value{
A sorted object of the same class as \code{lst}.
}
\description{
Sort an object \code{lst} by randomly rearranging it until it is sorted.
}
