% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_sort.R
\name{bubble_sort}
\alias{bubble_sort}
\title{bubble_sort}
\usage{
bubble_sort(lst=c(4,3,5,2,1), ascending = TRUE)
}
\arguments{
\item{lst}{An iterable object.}

\item{ascending}{Boolean specifying whether to sort in ascending or descending order.}
}
\value{
A sorted object of the same class as \code{lst}.
}
\description{
Sort an object \code{lst} using the bubble sort algorithm.
}
\references{
https://en.wikipedia.org/wiki/Bubble_sort
}
