% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sorted.R
\name{is_sorted}
\alias{is_sorted}
\title{is_sorted}
\usage{
is_sorted(lst, ascending = TRUE)
}
\arguments{
\item{lst}{An iterable object.}

\item{ascending}{Boolean specifying whether to check if \code{lst} is in ascending or descending order.}
}
\value{
Boolean, specifying whether \code{lst} is sorted.
}
\description{
Check if an iterable object \code{lst} is sorted.
}
